requirements.txt
PyYAML~=6.0.2
SQLAlchemy~=2.0.41
requests~=2.32.4
bs4~=0.0.2
beautifulsoup4~=4.13.4